+++ Client3/utility.py
@@ -12,6 +12,8 @@ import os
 import collections
 import glob
 import shutil
+import errno
+from socket import error as socket_error
 
 segmentDict = collections.defaultdict(list)
 lock = threading.Lock()
@@ -108,7 +110,7 @@ def process_data(threadName, delay, response, trackerFile, relevant_path, maxSeg
             #get latest updated tracker file
             listOfSegmentsInTrackerFile = parseTrackerFile(relevant_path + trackerFile)
 
-            #print "Peer 1: ", listOfSegmentsInTrackerFile,"\n\n"
+            #print "Peer 3: ", listOfSegmentsInTrackerFile,"\n\n"
             # get filename
             filename = listOfSegmentsInTrackerFile[len(listOfSegmentsInTrackerFile) - 1]
 
@@ -126,7 +128,7 @@ def process_data(threadName, delay, response, trackerFile, relevant_path, maxSeg
                 try:
                     #get latest updated tracker file
                     listOfSegmentsInTrackerFile = parseTrackerFile(relevant_path + trackerFile)
-                    #print " Peer 1: List of segments in tracker file: ",listOfSegmentsInTrackerFile,"\n\n"
+                    #print " Peer 3: List of segments in tracker file: ",listOfSegmentsInTrackerFile,"\n\n"
                     # get filename
                     filename = listOfSegmentsInTrackerFile[len(listOfSegmentsInTrackerFile) - 1]
                     listOfSegmentsInTrackerFile.pop()
@@ -137,14 +139,14 @@ def process_data(threadName, delay, response, trackerFile, relevant_path, maxSeg
                     for segmentLine in randomizedListOfSegments:
                         time.sleep(2)
                         # segmentLine = randomizedListOfSegments[index]
-                        #print "Peer 1: Try downloading current segment : from " ,  , "\n\n"
+                        #print "Peer 3: Try downloading current segment : from " ,  , "\n\n"
 
                         # calculate which segements to download, then download them
                         inf = segmentLine.split(":")
 
-                        #print "Peer 1: Try downloading segment from " , inf[2], " to ", inf[3], "\n\n"
+                        #print "Peer 3: Try downloading segment from " , inf[2], " to ", inf[3], "\n\n"
                         if int(inf[3]) - int(inf[2]) == int(fileSize):
-                            #print "Peer 1: Entire file is present with the peer ", "\n"
+                            #print "Peer 3: Entire file is present with the peer ", "\n"
                             #check if any segment in the entire file is pending to be downloaded in increasing order
                             startByte = int(inf[2])
                             endByte = startByte + int(maxSegmentSize)
@@ -167,7 +169,7 @@ def process_data(threadName, delay, response, trackerFile, relevant_path, maxSeg
                                     endByte = int(fileSize)
                         else:
                             isSegmentNeededToBeDownloaded = checkIfSegmentIsAlreadyDownloaded(filename, inf[2])
-                        #print " Peer 1: Does this segment need to be downloaded ? ", inf[2], " ", inf[3], isSegmentNeededToBeDownloaded, "\n\n"
+                        #print " Peer 3: Does this segment need to be downloaded ? ", inf[2], " ", inf[3], isSegmentNeededToBeDownloaded, "\n\n"
                         #downloadSegment_old(string)
                         
                         try:
@@ -181,14 +183,14 @@ def process_data(threadName, delay, response, trackerFile, relevant_path, maxSeg
                         # new download segment executes as thread, downloads segment indicated to stream, updates local tracker #
                         # pass arguments: open filestream, server ip, server port, segment begin, segment end
                 except:
-                    print " Peer 1 : try getting segments \n\n"
+                    print " Peer 3 : try getting segments \n\n"
 
                 if not os.path.exists(relevant_path+"temp/"):
-                    print " Peer 1: Complete file for this tracker file Downloaded - ", trackerFile, " \n\n" 
+                    print " Peer 3: Complete file for this tracker file Downloaded - ", trackerFile, " \n\n" 
 
                 else:
                     if os.path.isfile(fileNameTemp) == True:
-                        #print " Peer 1 : All segments are Downloaded: ", filename,"\n\n"
+                        #print " Peer 3 : All segments are Downloaded: ", filename,"\n\n"
                         fileNameTemp = mergeAllSegments(relevant_path, filename, fileNameTemp)
 
                         md5ForDownloadedFile = getMd5FromTrackerFile(relevant_path+trackerFile)
@@ -196,8 +198,9 @@ def process_data(threadName, delay, response, trackerFile, relevant_path, maxSeg
                         
                         if md5ForDownloadedFile.strip() == md5ForOriginalFile.strip():
                             #shutil.rmtree(relevant_path+"temp/")
-                            os.rename(fileNameTemp, relevant_path+filename)
-                            print "Peer 1 - File successfully Downloaded. Not Corrupted \n\n"
+                            if os.path.isfile(fileNameTemp) == True:
+                                os.rename(fileNameTemp, relevant_path+filename)
+                            print "Peer 3 - File successfully Downloaded. Not Corrupted \n\n"
 
                 #print "DONE\n\n"
                 
@@ -217,7 +220,7 @@ def createTrackerFile(filename, description, ip_address, PORT):
 
     #create the local copy of the tracker file
     #file = open("tracker"+str(timestamp)+".txt", "w")
-    string = "Peer 1: "+ "Create Tracker" + " Filename: "+actualFileName+" Filesize: "+ str(filesize)+" Description:"+description+" MD5:"+md5+" "+str(ip_address)+":"+str(PORT)+":0:"+str(filesize)+":"+str(timestamp)
+    string = "Peer 3: "+ "Create Tracker" + " Filename: "+actualFileName+" Filesize: "+ str(filesize)+" Description:"+description+" MD5:"+md5+" "+str(ip_address)+":"+str(PORT)+":0:"+str(filesize)+":"+str(timestamp)
     # print string, "\n\n"
     params = "GET command=createTracker&filename="+actualFileName+"&filesize="+str(filesize)+"&description="+description+"&md5="+md5+"&ip="+str(ip_address)+"&port="+str(PORT)+"&timestamp="+str(timestamp)
     #file.write(string)
@@ -232,7 +235,7 @@ def updateTrackerFile(filename, segmentLine):
     actualFileName = fileNameList[len(fileNameList) -1]
    
 
-    string = "Peer 1: "+" Updatetracker "+ " Filename: "+ actualFileName+ " start byte "+ segmentLine[2]+" End byte "+ segmentLine[3]+" ip-address "+ segmentLine[0]+" port "+segmentLine[1]
+    string = "Peer 3: "+" Updatetracker "+ " Filename: "+ actualFileName+ " start byte "+ segmentLine[2]+" End byte "+ segmentLine[3]+" ip-address "+ segmentLine[0]+" port "+segmentLine[1]
     # print string ,"\n\n"
     params = "GET command=updateTracker&filename="+actualFileName+"&s_byte="+segmentLine[2]+"&e_byte="+segmentLine[3]+"&ip="+segmentLine[0]+"&port="+segmentLine[1]+"&timestamp="+segmentLine[4]
     
@@ -276,7 +279,7 @@ def downloadSegment(threadName, fileNameTemp, server_addr, server_port, segment_
     socket1.connect((server_addr, int(server_port)))
     socket1.send(downloadSegmentStr)
     #data = socket1.recv(1024)
-    # print "Peer 1: Received data :" ,"\n"
+    # print "Peer 3: Received data :" ,"\n"
     lock.acquire()
     global file_to_write
     #with open(fileNameTemp, 'rb+') as file_to_write:   
@@ -302,7 +305,7 @@ def downloadSegment(threadName, fileNameTemp, server_addr, server_port, segment_
 
         with open(relevant_path+fileName+".track", "ab") as updateTrackerFileWithCurrentSegment:
             segmentLineStr =str(ip_address)+":"+str(peer_server_port)+":"+segment_beginaddr+":"+segment_endaddr+":"+str(int(time.time()))+"\n"
-            #print "Peer 1: Update tracker file with the current segment: \n"
+            #print "Peer 3: Update tracker file with the current segment: \n"
             # print segmentLineStr
             updateTrackerFileWithCurrentSegment.write(segmentLineStr)
         updateTrackerFileWithCurrentSegment.close()
@@ -316,15 +319,22 @@ def downloadSegment(threadName, fileNameTemp, server_addr, server_port, segment_
    
    
 def downloadSegmentInTempFolder(threadName, fileNameTemp, server_addr, server_port, segment_beginaddr, segment_endaddr, fileName, maxSegmentSize,ip_address,peer_server_port, relevant_path):
-    lock.acquire()
-    print "Try Downloading segment: ", " from " , segment_beginaddr, " to  ",segment_endaddr,"\n\n"
+    #lock.acquire()
+    print "\n\n Try Downloading segment: ", " from " , segment_beginaddr, " to  ",segment_endaddr,"\n\n"
     downloadSegmentStr = "download," + fileName + ","+segment_beginaddr+"," + segment_endaddr
     #print "Server Address : ", server_addr, " Server Port: ", server_port
     socket1 = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
     socket1.connect((server_addr, int(server_port)))
-    socket1.send(downloadSegmentStr)
-#data = socket1.recv(1024)
-    #print "Peer 1: Received data :" ,"\n" 
+    try:
+        socket1.send(downloadSegmentStr)
+
+    except socket_error as serr:
+        print " \n\n Peer 3 : The requesting Peer is terminated \n\n"
+        # if serr.errno != errno.ECONNREFUSED:
+        #     # Not the error we are looking for, re-raise
+        #     raise serr
+    #data = socket1.recv(1024)
+    #print "Peer 3: Received data :" ,"\n" 
     
     if not os.path.exists(relevant_path+"temp/"):
         os.makedirs(relevant_path+"temp/")
@@ -342,7 +352,7 @@ def downloadSegmentInTempFolder(threadName, fileNameTemp, server_addr, server_po
 
             with open(relevant_path+fileName+".track", "ab") as updateTrackerFileWithCurrentSegment:
                 segmentLineStr =str(ip_address)+":"+str(peer_server_port)+":"+segment_beginaddr+":"+segment_endaddr+":"+str(int(time.time()))+"\n"
-                #print "Peer 1: Update tracker file with the current segment: \n"
+                #print "Peer 3: Update tracker file with the current segment: \n"
                 #  print segmentLineStr
                 updateTrackerFileWithCurrentSegment.write(segmentLineStr)
             updateTrackerFileWithCurrentSegment.close()
@@ -353,7 +363,7 @@ def downloadSegmentInTempFolder(threadName, fileNameTemp, server_addr, server_po
 
     #print 'Client 1 : Download segment Successful from ',segment_beginaddr, " to ", segment_endaddr,"\n\n"
     socket1.close()
-    lock.release();
+    #lock.release();
 
 
 def updateDownloadedSegmentList(filename, s_byte):
@@ -418,9 +428,9 @@ def removeTrackerFilesForExistingFiles(relevant_path, allTrackerFilesList):
 
     if(len(toBeDownloadedList)>0):
         do="nothing"
-        # print " Peer 1 : To be downloaded List in : ", toBeDownloadedList
+        # print " Peer 3 : To be downloaded List in : ", toBeDownloadedList
     else:
-        print "Peer 1 : No new files that need to be downloaded. "
+        print "Peer 3 : No new files that need to be downloaded. "
     return toBeDownloadedList
 
 
@@ -436,9 +446,9 @@ def checkIfFileisDownloaded(relevant_path, trackerFile):
     trackerFileName = trackerFile.split(".track")
 
     if trackerFileName[0] in downloadedFilesList:
-        #print " Peer 1: Existing downloaded files : ", downloadedFilesList, " -  ", trackerFileName
+        #print " Peer 3: Existing downloaded files : ", downloadedFilesList, " -  ", trackerFileName
         return True
-    #print " Peer 1: Existing downloaded files : ", downloadedFilesList, " -  ", trackerFileName, ": ", trackerFileName[0]
+    #print " Peer 3: Existing downloaded files : ", downloadedFilesList, " -  ", trackerFileName, ": ", trackerFileName[0]
     return False
 
 def removeOriginallySharedFiles(relevant_path, allTrackerFilesList):
